+++ README.md
@@ -1,1 +1,4 @@
 # TelegramBot
+
+## Instalação
+```pip3 install -r requirements.txt```
