+++ handlers/start.py
@@ -1,0 +1,5 @@
+
+
+def start(update, context):
+    s = "hello, world!"
+    update.message.reply_text(s)
