+++ handlers/teste.py
@@ -1,0 +1,4 @@
+
+
+def teste(update, context):
+    print(update.message.chat_id)
